Other hack to handle the rendering of the posts in the paginated view
This is the problem: if we push the original post object in the pagination array, when it comes to render the pagination view jade is gonna render
first the pagination, which extends the base template, then when it comes to the post it's gonna render the post as it's own
page extending the base template as well, this means that we're gonna end up with a weird page inside the page.

I'm sure that to avoid it there must be some option to pass to the jade compiler, but i haven't find it yet.